requirements.txt
celery
redis
faster-whisper
streamlit
yt-dlp
transformers
torch
opencv-python
numpy
Pillow
scenedetect
fastapi
uvicorn
langchain
langchain-openai
langchain-anthropic
requests
llama-index-core
llama-index-embeddings-huggingface
llama-index-llms-anthropic
sentence-transformers
python-dotenv
